Paraphrase transcript of 2012-12-05 for improved readability
Chips out the transcripts from the codeblock and formats them
as normal question and answer type lines, with a few spelling and
missing word fixes.

The command history and transcripts don't quite follow a trackline
so it couldn't hurt to bundle them up with each other's day half
respectively instead of having them split.

Expecting to make the Q&A transcript easier to read overall.